(Kyle & Stewart) Started a new version, started trying to get api data
Working off of the project 4 template was turning into a mess (everything kept breaking, too much code to change around). We instead started working from the project 3 template so we could just try to get some api data flowing in the app. It should work, but for some reason our json parsing isn't working. If that is fixed we should be in a good place